feat: Migrate streamlit demo to a new multiturn chatbot implementation with enhanced message handling and history management
